requirements.txt
pyodbc==5.0.1
google-cloud-bigquery==3.14.1
pandas==2.1.4
sentry-sdk==1.39.2
python-dotenv==1.0.0
pyarrow>=10.0.0
